02_OCL/requirements.txt
beautifulsoup4
Crawl4AI
html2text
lancedb
openai
playwright
pypdf
python-dotenv
requests
tiktoken
